requirements.txt
Flask>=2.3.0
Flask-Login>=0.6.2
Flask-SQLAlchemy>=3.0.0
python-dotenv>=1.0.0
gunicorn>=20.1.0
Flask-Migrate>=4.0.0
Flask-Talisman>=1.0.0
Flask-WTF>=1.1.1
Flask-Limiter>=2.9.1
psycopg2-binary>=2.9.7
boto3>=1.26.0
